refactor: move mutation functions to Channels Db mod
