+++ lib/ae_mdw/channels.ex
@@ -4,16 +4,10 @@ defmodule AeMdw.Channels do
   """
 
   alias :aeser_api_encoder, as: Enc
-  alias AeMdw.Blocks
   alias AeMdw.Collection
-  alias AeMdw.Db.ChannelCloseMutation
-  alias AeMdw.Db.ChannelOpenMutation
-  alias AeMdw.Db.ChannelSpendMutation
-  alias AeMdw.Db.ChannelUpdateMutation
   alias AeMdw.Db.Model
-  alias AeMdw.Db.Mutation
   alias AeMdw.Db.State
-  alias AeMdw.Node
+  alias AeMdw.Error.Input, as: ErrInput
   alias AeMdw.Txs
   alias AeMdw.Util
   alias AeMdw.Validate
@@ -23,8 +17,7 @@ defmodule AeMdw.Channels do
   @typep state() :: State.t()
   @typep pagination() :: Collection.direction_limit()
   @typep range() :: {:gen, Range.t()} | nil
-  @typep cursor() :: binary() | nil
-  @typep bi_txi() :: Blocks.bi_txi()
+  @typep cursor() :: Collection.pagination_cursor()
 
   @type channel() :: map()
 
@@ -32,119 +25,20 @@ defmodule AeMdw.Channels do
   @table_activation Model.ActiveChannelActivation
 
   @spec fetch_active_channels(state(), pagination(), range(), cursor()) ::
-          {cursor(), [channel()], cursor()}
+          {:ok, cursor(), [channel()], cursor()}
   def fetch_active_channels(state, pagination, range, cursor) do
-    cursor = deserialize_cursor(cursor)
-    scope = deserialize_scope(range)
+    with {:ok, cursor} <- deserialize_cursor(cursor) do
+      scope = deserialize_scope(range)
 
-    {prev_cursor, expiration_keys, next_cursor} =
-      state
-      |> build_streamer(scope, cursor)
-      |> Collection.paginate(pagination)
+      {prev_cursor, expiration_keys, next_cursor} =
+        state
+        |> build_streamer(scope, cursor)
+        |> Collection.paginate(pagination)
 
-    channels = render_list(state, expiration_keys)
+      channels = render_list(state, expiration_keys)
 
-    {serialize_cursor(prev_cursor), channels, serialize_cursor(next_cursor)}
-  end
-
-  @spec close_mutual_mutations(bi_txi(), Node.aetx()) :: [Mutation.t()]
-  def close_mutual_mutations(bi_txi, tx) do
-    channel_pk = :aesc_close_mutual_tx.channel_pubkey(tx)
-
-    release_amount =
-      :aesc_close_mutual_tx.initiator_amount_final(tx) +
-        :aesc_close_mutual_tx.responder_amount_final(tx)
-
-    [
-      ChannelCloseMutation.new(channel_pk, bi_txi, release_amount)
-    ]
-  end
-
-  @spec close_solo_mutations(bi_txi(), Node.aetx()) :: [Mutation.t()]
-  def close_solo_mutations(bi_txi, tx) do
-    channel_pk = :aesc_close_solo_tx.channel_pubkey(tx)
-
-    [
-      ChannelUpdateMutation.new(channel_pk, bi_txi)
-    ]
-  end
-
-  @spec set_delegates_mutations(bi_txi(), Node.aetx()) :: [Mutation.t()]
-  def set_delegates_mutations(bi_txi, tx) do
-    channel_pk = :aesc_set_delegates_tx.channel_pubkey(tx)
-
-    [
-      ChannelUpdateMutation.new(channel_pk, bi_txi)
-    ]
-  end
-
-  @spec close_slash_mutations(bi_txi(), Node.aetx()) :: [Mutation.t()]
-  def close_slash_mutations(bi_txi, tx) do
-    channel_pk = :aesc_slash_tx.channel_pubkey(tx)
-
-    [
-      ChannelUpdateMutation.new(channel_pk, bi_txi)
-    ]
-  end
-
-  @spec force_progress_mutations(bi_txi(), Node.aetx()) :: [Mutation.t()]
-  def force_progress_mutations(bi_txi, tx) do
-    channel_pk = :aesc_force_progress_tx.channel_pubkey(tx)
-
-    [
-      ChannelUpdateMutation.new(channel_pk, bi_txi)
-    ]
-  end
-
-  @spec slash_mutations(bi_txi(), Node.aetx()) :: [Mutation.t()]
-  def slash_mutations(bi_txi, tx) do
-    channel_pk = :aesc_slash_tx.channel_pubkey(tx)
-
-    [
-      ChannelUpdateMutation.new(channel_pk, bi_txi)
-    ]
-  end
-
-  @spec offchain_mutations(bi_txi(), Node.aetx()) :: [Mutation.t()]
-  def offchain_mutations(bi_txi, tx) do
-    channel_pk = :aesc_offchain_tx.channel_pubkey(tx)
-
-    [
-      ChannelUpdateMutation.new(channel_pk, bi_txi)
-    ]
-  end
-
-  @spec settle_mutations(bi_txi(), Node.aetx()) :: [Mutation.t()]
-  def settle_mutations(bi_txi, tx) do
-    channel_pk = :aesc_close_settle_tx.channel_pubkey(tx)
-
-    %{"initiator_amount_final" => initiator_amount, "responder_amount_final" => responder_amount} =
-      :aesc_settle_tx.for_client(tx)
-
-    [
-      ChannelCloseMutation.new(channel_pk, bi_txi, initiator_amount + responder_amount)
-    ]
-  end
-
-  @spec open_mutations(bi_txi(), Node.aetx()) :: [Mutation.t()]
-  def open_mutations(bi_txi, tx), do: [ChannelOpenMutation.new(bi_txi, tx)]
-
-  @spec deposit_mutations(bi_txi(), Node.aetx()) :: [Mutation.t()]
-  def deposit_mutations(bi_txi, tx) do
-    channel_pk = :aesc_deposit_tx.channel_pubkey(tx)
-
-    [
-      ChannelSpendMutation.new(channel_pk, bi_txi, :aesc_deposit_tx.amount(tx))
-    ]
-  end
-
-  @spec withdraw_mutations(bi_txi(), Node.aetx()) :: [Mutation.t()]
-  def withdraw_mutations(bi_txi, tx) do
-    channel_pk = :aesc_withdraw_tx.channel_pubkey(tx)
-
-    [
-      ChannelSpendMutation.new(channel_pk, bi_txi, -:aesc_withdraw_tx.amount(tx))
-    ]
+      {:ok, serialize_cursor(prev_cursor), channels, serialize_cursor(next_cursor)}
+    end
   end
 
   @spec channels_opened_count(state(), Txs.txi(), Txs.txi()) :: non_neg_integer()
@@ -204,9 +98,9 @@ defmodule AeMdw.Channels do
     with [exp_height, channel_pk] <-
            Regex.run(~r/\A(\d+)-(\w+)\z/, cursor_bin, capture: :all_but_first),
          {:ok, channel_pk} <- Validate.id(channel_pk, [:channel]) do
-      {String.to_integer(exp_height), channel_pk}
+      {:ok, {String.to_integer(exp_height), channel_pk}}
     else
-      _invalid -> nil
+      _invalid -> ErrInput.Cursor.exception(value: cursor_bin)
     end
   end
